I FIXED THE DAMN BUG WHERE THE MAIN MENU BROKE AFTER QUITTING TO THE MAIN MENU YAY.
It was a simple fix too! I just had to get the game to unpause after clicking the `QuitToMainMenuButton` button.

I also updated the main menu to reflect this change and replaced all instances of `return` with `pass` statements in `Stage.gd` for better style.

Oh, and I "added" a rematch button to `WinMenu.tscn` (that was originally hidden and I gave it logic to reload the root scene).